requirements.txt
# VSHORAD Aircraft Detection System
# Core dependencies

# Detection
ultralytics>=8.3.0

# Classification
timm>=1.0.11
albumentations>=1.4.18

# Core ML
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0,<2.0.0

# Tracking (included in ultralytics)
# bytetrack via ultralytics

# Utilities
opencv-python>=4.8.0
PyYAML>=6.0
tqdm>=4.65.0
scikit-learn>=1.3.0

# Optional: TensorRT export (for Embedded tier)
# onnx>=1.16.0
# onnxruntime-gpu>=1.18.0
# tensorrt (system-specific)

# Optional: Training
# codecarbon>=2.3.0
# pandas>=2.0.0
# matplotlib>=3.7.0
# seaborn>=0.12.0
